refactor: restructure NblmClient and modularize HTTP handling
- Removed the existing client.rs file and split its functionality into multiple modules for better organization.
- Introduced a new HttpClient struct to encapsulate HTTP request logic, improving separation of concerns.
- Added a UrlBuilder for constructing API endpoint URLs, enhancing clarity and maintainability.
- Implemented a Retryer for handling retry logic on failed requests, improving robustness.
- Updated NblmClient to utilize the new HttpClient and UrlBuilder, streamlining its interface.
- Added API endpoint implementations for audio, notebooks, and sources, enhancing functionality.
- Introduced unit tests for new modules to ensure reliability and correctness.
