+++ crates/nblm-core/src/client/api/mod.rs
@@ -1,0 +1,4 @@
+// API endpoint implementations
+mod audio;
+mod notebooks;
+mod sources;
