+++ crates/nblm-core/src/lib.rs
@@ -1,13 +1,11 @@
 pub mod auth;
-mod client;
+pub mod client;
 mod error;
 pub mod models;
-mod retry;
 
 pub use auth::{EnvTokenProvider, GcloudTokenProvider, StaticTokenProvider, TokenProvider};
-pub use client::NblmClient;
+pub use client::{NblmClient, RetryConfig, Retryer};
 pub use error::{Error, Result};
-pub use retry::{RetryConfig, Retryer};
 
 use std::sync::Arc;
